+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/de/1.0.0/).
 
+## [Unreleased]
+### Added
+- SQL-Entry-Field
+
 ## [v1.2.4]
 ### Added
 - Able to delete person